Improving detail in Kaiko intermediate output
We now have a || scan | peptide | protein | taxa | taxa stats || table for finer control of the logic which generates the FASTA file.

Also added Kaiko_parse_uniref.py, which parses the uniref100 database FASTA file and creates and txt file index with the position of every protein, and a txt file containing the taxa proteome size for each of the 600000+ individual taxa within uniref100, allowing the us to filter based on taxa size, as Joon had suggested.