Updated reading of mgf files
The old method was dependent on a very specific order and number of header variables in the mgf spectrum. Now the new method accepts variable lengths, though it requires PEPMASS and CHARGE, scan can be inferred from the title when missing.

Needed to read the mgf files from Kansas soil